ml/requirements.txt
scikit-learn>=1.4.0
lightgbm>=4.3.0
mlflow>=2.10.0
psycopg2-binary>=2.9.9
pandas>=2.2.0
